Fix lines with false-* applications
Sometimes tests intentionally not await to test warnings.
It's fine if we "over-apply" in these cases since that's not something
that should be done in userspace.

In other places it's an `act` that's received by `expect`.
We could codemod these but they're rare enough
that we may not need to do this.