[DevTools] Always attempt to mount dehydrated roots
`mountFiberRecursively` bails out of hydrating dehydrated roots nowadays.
This makes the check in `handleCommitFiberRoot` redundant.

Mostly doing this because the check in `handleCommitFiberRoot`
wasn't mirrored in the initial mount path.
`handleCommitFiberRoot` always attempted to mount dehydrated
roots when it first encountered a HostRoot.
But on updates, it treated a previously dehydrated HostRoot
as newly mounted.
This resulted in recording mounts twice which is not supported.